+++ result/result.py
@@ -52,22 +52,6 @@ class Result(object):
         """
         return self._val if self.is_err() else None
 
-    def and_then(self, function, **kwargs):
-        """
-        Return the error if this is an `Err` type. Return `function` calls return value otherwise.
-        """
-        if self._type == 'error':
-            return self
-        return function(self._val, **kwargs)
-
-    def or_else(self, function, **kwargs):
-        """
-        Return the error if this is an `Ok` type. Return `function` calls return value otherwise.
-        """
-        if self._type == 'ok':
-            return self
-        return function(self._val, **kwargs)
-
     @property
     def value(self):
         """