+++ result/tests.py
@@ -58,24 +58,6 @@ def test_err_method():
     assert n.err() == 'nay'
 
 
-def test_and_ten():
-    def sq(x):
-        return Ok(x * x)
-
-    assert Ok(2).and_then(sq) == Ok(4)
-    assert Err(2).and_then(sq) == Err(2)
-    assert Ok(2).and_then(lambda x, y, z: Ok(x * y * z), y=2, z=4) == Ok(16)
-
-
-def test_or_else():
-    def sq(x):
-        return Ok(x * x)
-
-    assert Ok(2).or_else(sq) == Ok(2)
-    assert Err(2).or_else(sq) == Ok(4)
-    assert Err(2).or_else(lambda x, y, z: Ok(x * y * z), y=2, z=4) == Ok(16)
-
-
 def test_no_constructor():
     """
     Constructor should not be used directly.