+++ CHANGELOG.md
@@ -13,7 +13,11 @@ Possible log types:
 
 ## [Unreleased]
 
- - [added] __eq__ magic method
+ - ...
+
+## [0.2.2] - 2016-09-21
+
+ - [added] `__eq__` magic method
 
 ## [0.2.0] - 2016-05-05
 
@@ -27,7 +31,8 @@ Possible log types:
 
  - Initial version
 
-[Unreleased]: https://github.com/dbrgn/result/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/dbrgn/result/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/dbrgn/result/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/dbrgn/result/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/dbrgn/result/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/dbrgn/result/compare/3ca7d83...v0.1.0